update nginx config, update docker compose, fix multiple import errors, update requirements, add Decimal type to some tables

+++ backend/app/database/models/residue.py
@@ -1,4 +1,5 @@
 from typing import TYPE_CHECKING
+from decimal import Decimal
 
 from sqlmodel import Field, SQLModel, Relationship
 
@@ -9,8 +10,8 @@ if TYPE_CHECKING:
 class ResidueBase(SQLModel):
     name: str
     charge: int = Field(index=True)
-    hodropathy: float = Field(index=True)
-    polarity: float = Field(index=True)
+    hodropathy: Decimal = Field(index=True, decimal_places=3)
+    polarity: Decimal = Field(index=True, decimal_places=3)
     mutability: int = Field(index=True)
